Additional error checking
Noticed some tooltips within the retail client are missing the item name data and also contain an item link without the item ID data. Unable to detect the item without this data so I introduced an error condition during the item link parse.
This occurs within the right side panel of the professions dialog, when selecting an item to craft.